refactoring folder structure + other changes
